sl1: Support 'Fade' and 'Slow' bottom exposure types
Summary:
- Refs #3
